app/requirements.txt
django==3.2.5
djangorestframework==3.13.1
psycopg2-binary==2.9.3
django-celery-results==2.3.1 
celery[redis]==5.2.6
biopython==1.78
